Added spsa wrapper
Use just like TNOptimizer, but with SPSA method. 

It's inefficient (e.g. calculating grads when we don't need to) and uses two function calls per step, but it works fine for the circuits we care about. (Can optimize if needed)

results are stored in a new results_dict in the object